# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hebrew RTL E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      parameters:
        - name: parentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ──────────────────────────────────────────────────────────────
...
  /comparison:
    get:
      operationId: getComparison
      tags: [comparison]
      summary: Get comparison list
      responses:
        "200":
          description: Comparison products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: addToComparison
      tags: [comparison]
      summary: Add product to comparison
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonActionBody"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: clearComparison
      tags: [comparison]
      summary: Clear comparison list
      responses:
        "204":
          description: Cleared
  /comparison/{productId}:
    delete:
      operationId: removeFromComparison
      tags: [comparison]
      summary: Remove product from comparison
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── RECENTLY VIEWED ────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: active
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResult"

  # ─── INVENTORY ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success

    # ── Category ──
    Category:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - slug
        - sortOrder
        - isActive
        - createdAt
        - updatedAt
    CreateCategoryBody:
      type: object
      properties:
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
      required:
        - nameHe
        - slug

    # ── Brand ──
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        descriptionHe:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties: true
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        viewsCount:
          type: integer
        salesCount:
          type: integer
        ratingAverage:
          type: number
        ratingCount:
          type: integer
        weight:
          type: ["number", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - slug
        - price
        - stockQuantity
        - isActive
        - isFeatured
        - viewsCount
        - salesCount
        - ratingAverage
        - ratingCount
        - images
        - tags
        - createdAt
        - updatedAt
...
    ComparisonActionBody:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId

    # ── Recently Viewed ──
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed, free_shipping]
        value:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxDiscountAmount:
          type: ["number", "null"]
        applicableCategories:
          type: array
          items:
            type: integer
        applicableBrands:
          type: array
          items:
            type: integer
        usageLimit:
          type: ["integer", "null"]
        usageLimitPerUser:
          type: ["integer", "null"]
        usedCount:
          type: integer
        isActive:
          type: boolean
        isStackable:
          type: boolean
        startsAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - type
        - value
        - usedCount
        - isActive
        - isStackable
        - applicableCategories
        - applicableBrands
        - createdAt
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed, free_shipping]
        value:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxDiscountAmount:
          type: ["number", "null"]
        applicableCategories:
          type: array
          items:
            type: integer
        applicableBrands:
          type: array
          items:
            type: integer
        usageLimit:
          type: ["integer", "null"]
        usageLimitPerUser:
          type: ["integer", "null"]
        isActive:
          type: boolean
        isStackable:
          type: boolean
        startsAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - code
        - type
        - value
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        cartTotal:
          type: number
        userId:
          type: ["integer", "null"]
      required:
        - code
        - cartTotal
    CouponValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
        discountAmount:
          type: number
        message:
          type: ["string", "null"]
      required:
        - valid
        - discountAmount

    # ── Inventory ──
...